[#29] Refactor and add tests for childProcess helper
